net: be more aggresive wrt removing from known peers DB
Prior to this commit we end up trying to reconnect to peers every time
we start up. Be more aggressive in removing bad peers from the DB list
of known peers.

Also make sure to re-add the seed node every time we start up, if we
have an empty list of known peers.