multi: tracing in CLI, request ID handling
Add capabilities for setting request IDs through header values, which we
include in tracing spans. We then set randomly generated request IDs in
the CLI, such that it's possible to trace requests/responses going
through the system.

This is conceptually very similar to https://github.com/LayerTwo-Labs/bip300301_enforcer/pull/150.

One would think that we could do this with the actual `id` field of the
JSON-RPC bodies. This is not possible, because the `jsonrpsee` library
does not make it possible to configure this.

Upstream issue for this https://github.com/paritytech/jsonrpsee/issues/1524